[BROOKLYN-183] Load brooklyn.properties on OSGi init
