+++ karaf/apache-brooklyn/src/main/resources/etc/org.apache.brooklyn.osgilauncher.cfg
@@ -24,6 +24,12 @@
 # Most of these will migrate to a permanent cfg file, once the components themselves get refactored
 # and the configuration options are agreed upon
 
+# Location of the global brooklyn.properties file
+#globalBrooklynPropertiesFile=~/.brooklyn/brooklyn.properties
+
+# Location of the local brooklyn.properties file, normally specified at the cli. Overrides properties from the global set.
+#localBrooklynPropertiesFile=
+
 
 # Ignore catalog subsystem failures during startup (default is to continue, so errors can be viewed via the API)
 #ignoreCatalogErrors=true
