+++ karaf/features/src/main/feature/feature.xml
@@ -286,4 +286,10 @@
         <bundle dependency="true">mvn:com.google.guava/guava/${guava.version}</bundle>
     </feature>
 
+    <feature name="brooklyn-osgi-launcher" version="${project.version}" description="Brooklyn init">
+        <feature>brooklyn-core</feature>
+        <bundle>mvn:org.apache.brooklyn/brooklyn-launcher-common/${project.version}</bundle>
+        <bundle>mvn:org.apache.brooklyn/brooklyn-karaf-init/${project.version}</bundle>
+    </feature>
+
 </features>
