+++ karaf/pom.xml
@@ -53,6 +53,7 @@
   </properties>
 
   <modules>
+    <module>init</module>
     <module>features</module>
     <module>apache-brooklyn</module>
     <module>commands</module>
